fixes for pull request
